Added sorting functionality to the Staff and Customers class. Added a filter to the Staff and Customer index and controller.

+++ src/FinalTeamProject/Controllers/CustomersController.cs
@@ -16,13 +16,35 @@ namespace FinalTeamProject.Controllers
 
         public CustomersController(BookingContext context)
         {
-            _context = context;    
+            _context = context;
         }
 
         // GET: Customers
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string sortOrder, string searchString)
         {
-            return View(await _context.Customers.ToListAsync());
+            ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+            ViewData["CurrentFilter"] = searchString;
+
+            var customers = from s in _context.Customers
+                           select s;
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                customers = customers.Where(s => s.LastName.Contains(searchString)
+                                       || s.FirstName.Contains(searchString));
+            }
+
+            switch (sortOrder)
+            {
+                case "name_desc":
+                    customers = customers.OrderByDescending(s => s.LastName);
+                    customers = customers.OrderByDescending(s => s.FirstName);
+                    break;
+                default:
+                    customers = customers.OrderBy(s => s.LastName);
+                    customers = customers.OrderBy(s => s.FirstName);
+                    break;
+            }
+            return View(await customers.AsNoTracking().ToListAsync());
         }
 
         // GET: Customers/Details/5
@@ -58,13 +80,24 @@ namespace FinalTeamProject.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,FirstName,LastName,Telephone")] Customer customer)
+        public async Task<IActionResult> Create
+            ([Bind("ID,FirstName,LastName,Telephone")] Customer customer)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(customer);
-                await _context.SaveChangesAsync();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    _context.Add(customer);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
             }
             return View(customer);
         }
@@ -88,50 +121,59 @@ namespace FinalTeamProject.Controllers
         // POST: Customers/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+
+
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,FirstName,LastName,Telephone")] Customer customer)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != customer.ID)
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var customerToUpdate = await _context.Customers.SingleOrDefaultAsync(s => s.ID == id);
+            if (await TryUpdateModelAsync<Customer>(
+                customerToUpdate,
+                "",
+                s => s.FirstName, s => s.LastName, s => s.Telephone))
             {
                 try
                 {
-                    _context.Update(customer);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction("Index");
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!CustomerExists(customer.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists, " +
+                        "see your system administrator.");
                 }
-                return RedirectToAction("Index");
             }
-            return View(customer);
+            return View(customerToUpdate);
         }
 
         // GET: Customers/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var customer = await _context.Customers.SingleOrDefaultAsync(m => m.ID == id);
+            var customer = await _context.Customers
+                .AsNoTracking()
+                .SingleOrDefaultAsync(m => m.ID == id);
             if (customer == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(customer);
@@ -142,15 +184,25 @@ namespace FinalTeamProject.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var customer = await _context.Customers.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Customers.Remove(customer);
-            await _context.SaveChangesAsync();
-            return RedirectToAction("Index");
-        }
+            var customer = await _context.Customers
+                .AsNoTracking()
+                .SingleOrDefaultAsync(m => m.ID == id);
+            if (customer == null)
+            {
+                return RedirectToAction("Index");
+            }
 
-        private bool CustomerExists(int id)
-        {
-            return _context.Customers.Any(e => e.ID == id);
+            try
+            {
+                _context.Customers.Remove(customer);
+                await _context.SaveChangesAsync();
+                return RedirectToAction("Index");
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction("Delete", new { id = id, saveChangesError = true });
+            }
         }
     }
 }
